+++ toolkit/template/scaffold/project/cpp/qt5/example/qml/main.cpp
@@ -9,6 +9,8 @@ int main(int argc, char *argv[]) {
 
   QGuiApplication app(argc, argv);
 
+  QGuiApplication::setFont(QFont("Microsoft YaHei", 14));
+
   Core *core = new Core(&app);
 
   QQmlApplicationEngine engine;