+++ toolkit/template/scaffold/project/cpp/qt5/example/qml/main.qml
@@ -1,5 +1,9 @@
 import QtQuick 2.12
 import QtQuick.Window 2.12
+import QtQuick.Controls 2.12
+import QtQuick.Dialogs 1.1
+import QtQuick.Layouts 1.1
+import Qt.labs.platform 1.1
 
 import "main.js" as MainJS
 
@@ -8,7 +12,7 @@ Window {
     visible: true
     width: 640
     height: 480
-    title: qsTr("Download")
+    title: qsTr("{{ project_slug.words_capitalized }}")
 
     Component.onCompleted: {
         MainJS.init(window, core)